pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoharness"
version = "0.1.0"
description = "Behavioral governance middleware for AI agents with multi-tier pipeline modes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AutoHarness Contributors" },
]
keywords = [
    "ai",
    "agent",
    "governance",
    "safety",
    "middleware",
    "llm",
    "claude",
    "openai",
    "security",
    "audit",
    "autoharness-engineering",
    "context-engineering",
    "tool-governance",
    "anti-distillation",
    "model-routing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30.0"]
openai = ["openai>=1.0.0"]
langchain = ["langchain-core>=0.2.0"]
all = [
    "anthropic>=0.30.0",
    "openai>=1.0.0",
    "langchain-core>=0.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
autoharness = "autoharness.cli.main:cli"

[project.urls]
Homepage = "https://github.com/aiming-lab/AutoHarness"
Documentation = "https://autoharness.dev"
Repository = "https://github.com/aiming-lab/AutoHarness"
Issues = "https://github.com/aiming-lab/AutoHarness/issues"

[tool.hatch.build.targets.wheel]
packages = ["autoharness"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
